added delete API's

+++ OnlineRailwayReservation/src/main/java/com/sunbeam/services/UserService.java
@@ -12,7 +12,7 @@ public interface UserService {
 	User findByEmail(String email);
 	List<User>findAll();
 	User save(User user);
-	Boolean deleteById(int id);
+	Boolean deleteById(Long id);
 	User authenticate(String email, String password );
 	User update(User user);
 	User findById(int id);
